+++ Modules/Microsoft365DSC/Modules/M365DSCUtil.psm1
@@ -1403,7 +1403,12 @@ function Assert-M365DSCBlueprint
 
         [Parameter()]
         [Switch]
-        $Parallel
+        $Parallel,
+
+        [Parameter()]
+        [ValidateRange(1, [System.Int32]::MaxValue)]
+        [System.Int32]
+        $ThrottleLimit = 5
     )
 
     #Ensure the proper dependencies are installed in the current environment.
@@ -1487,9 +1492,17 @@ function Assert-M365DSCBlueprint
         Write-Host "Selected BluePrint contains ($($ResourcesInBluePrint.Length)) components to assess."
 
         # Call the Export-M365DSCConfiguration cmdlet to extract only the resource
-        # types contained within the BluePrint;
+        # types contained within the BluePrint
         Write-Host "Initiating the Export of those ($($ResourcesInBluePrint.Length)) components from the tenant..."
         $TempExportName = 'TempExport_' + (New-Guid).ToString() + '.ps1'
+
+        $additionalParams = @{}
+        if ($Parallel.IsPresent)
+        {
+            $additionalParams = @{
+                ThrottleLimit = $ThrottleLimit
+            }
+        }
         Export-M365DSCConfiguration -Components $ResourcesInBluePrint `
             -Path $env:TEMP `
             -FileName $TempExportName `
@@ -1502,7 +1515,7 @@ function Assert-M365DSCBlueprint
             -CertificatePath $CertificatePath `
             -CertificatePassword $CertificatePassword `
             -ManagedIdentity:$ManagedIdentity.IsPresent `
-            -AccessTokens $AccessTokens
+            -AccessTokens $AccessTokens @additionalParams
 
         # Call the New-M365DSCDeltaReport configuration to generate the Delta Report between
         # the BluePrint and the extracted resources;